Handle nested source directories
The Template.data.mk module did not handle sub-directory files to
install correctly, becoming confused as to which directories had to be
created.

Ironically this failed on zmk itself :)

This patch fixes the issue by associating a new instance attribute,
target_dir, that contains the full directory path (excluding DESTDIR) to
which the file needs to be copied to.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>