Make .version authoritative, generate .version-from-git
This patch performs the following trick. The .version-from-git file can
be generated if we have a git directory in the source path. Such file is
automatically renamed when making distribution tarballs, to .version.
When .version is picked up in the source tree it is ultimately
authoritative, overriding (and in fact, not even considering) the
version derived from the git repository

This allows downstream packaging to use git and not confuse our system.

Fixes: #2
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>