Detect CI with partial history
CI systems are incentivised to use the fewest amount of time, disk space
and network bandwidth and will use shallow git checkouts whenever
possible.

When VERSION_git is empty despite all other things looking correct AND
when the CI variable is not undefined, produce an explicit error with a
helpful message.

Fixes: #25
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>