Simplify C/C++ compiler de-generalization
Operating systems install the compiler stack under familiar names such
as "cc" for the C compiler and "c++" or "g++" for the C++ compiler. That
binary is often a symbolic link to the actual compiler binary, which
contains the full name, either Gcc or Clang.

This patch simplifies this logic.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>