Rename MATCH to GREP
The original name was derived from a similar test helper available in
the spread program. It is not a good name because it clashes with spread
and has different semantics. GREP is really a wrapper around grep,
without any additional options. One can use GREP with any options
desired while MATCH comes with pre-baked behavior.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>