Rewrite and fix DESTDIR handling
DESTDIR handling was buggy. To ensure it's not broken add all the tests
for most commonly used modules. Modules such as Header and ManPage were
reduced to thin shim around existing modules to reduce complexity.

The Directories (plural) module is no longer handling DESTDIR. This
responsibility is now shifted to the Directory module/template. Lastly
the Symlink module was fixed as well, and in addition made use AllClean
directly.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>