Fix bundling and removal logic
ZMK should be able to bundle itself into a project and remove itself
when dist-cleaning but only when it is provided externally. In addition
the case of zmk self-bundling inside its own project needs to behave
correctly.

The main change is that $(ZMK.Path) no longer includes a trailing slash
and can now be compared against srcdir $(reliably). This removes the
need to special case the zmk project name.

The important property is that "make distclean" no longer removes the
configure script, as typical packaging helpers will be unable to
re-generate the bundled ZMK installation.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>