Fix InstallUninstall handling of relative paths
InstallUninstall no longer reuses the relative directory in the source
tree to represent the target directory. Instead the InstallDir property
is the only way to influence the final path. This was a late discovery
but one that is ultimately only sensible. ZMK itself was adjusted to
install correctly in this new scheme.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>