Allow gcc/g++ to be uninstalled
When a configured build which is using GCC (for either C or C++) is
testing to see if it is a cross-compiler, it relies on the output of
"gcc -dumpmachine" and the output of "$(CC) -dumpmachine)" to know if
cross-compiling is in effect. This assumes that gcc and g++ are
installed, which may not be the case.

If gcc or g++ are absent, the cross compiler check will no longer
produce a warning such as:

    make: sparc64-linux-gnu-g++: Command not found

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>