Improve srcdir handling in configure script
When the configure script is newer than the configuration file then the
rule would re-execute the configuration script but along the way the
customized srcdir would get lost.

There are several changes bundled here as well. The parsing section of
the script no longer runs in a sub-shell with redirection to the
configuration file. Instead only after all of parsing, which now also
handles srcdir=, a sub-shell writes to the configuration file. This
simplifies correct handling of srcdir.

In addition OBJCXXFLAGS is now handled and stored.

In addition, the generated comments now explain the default value
of uncustomized property.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>