Rewrite and fix Tarball.mk
There are several changes that are bundled here.
All the variables are now namespaced. Several things are defined
as private variables that can be overridden for testing. GNU vs BSD
tar is supported much better now, as is MacOS specific filtering.
Lastly tarballs now work correctly when used out-of-tree.

A tiny new feature is the option to never bundle files into the generated
archive. This is also related to the change in how bundling is performed,
removing the need to put a copy of zmk into the source tree.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>