Fix parsing of wc -l on MacOS
Eh, portability. The output of "wc -l" on MacOS doesn't look the same
way as it does on Linux, so the trick with "cut -d ' ' -f 1" does not
function correctly.

This was not caught in CI as it only manifests itself in the failure
code path and only in a user-interface sort of way.

To work around this issue, use wc -l with re-directed stdin, and have
the excess spaces parsed by bash during the number coercion in the test
comparison.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>