Introduce Project.{Name,Version}
Those variables have the same purpose as NAME and VERSION, and in fact
are more generic version of the two. While both NAME and VERSION default
to $(error), the new variables show either NAME and VERSION if defined
or expand to an empty value.

This allows safer usage of Project.Name instead of NAME and
Project.Version instead of VERSION, without having to carefully handle
the error defaults.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>