Fix support for sub-directories in ObjectGroup
When a project uses non-flat layout, for example with sources in src/
and public header files in include/, zmk would misunderstand this and
try to create files in non-existent directories.

Now all objects know which source directory they come from, and will
build correctly, including out-of-tree builds.

In addition, since now it is easy to do so, add support for alternative
c++ extensions: .cc and .cxx, this is in addition to the .cpp extension
that was supported before.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>