Support unclean directory names
When the Directory template is expanded with an unclean path, such as
/path/to/foo/, the trailing slash is automatically removed and the mkdir
rule exists only of the canonical path: /path/to/foo. This is correct
but may lead to misleading results when some other rule depends on the
unclean path's existence. For simplicity emit a support rule which
order-depends on the clean path from the unclean path.

In terms of make rules it looks somewhat like this:

/path/to/foo:
	install -d $@
/path/to/foo/: | /path/to/foo

In particular, this fixes sub-directory creation, during out-of-tree
builds of sources using src/foo.c layout.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>