Rename makefile configuration to config.$(NAME).mk
This moves the file from GNUmakefile.$(NAME).configured.mk to
config.$(NAME).mk - if we have to do a breaking change then let's at
least use a name that will match upcoming config.$(NAME).h.

The backstory of the first rename is that the GNUmakefile.configured.mk
caused unexpected behavior when zmk's own configuration file was
included instead of the per-project makefile (which did not exist at the
time). The second rename is to align the created config file with the
upcoming probe system which will generate config.$(NAME).h. It's nicer
when those files have matching names.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>