Fix out-of-tree build with read-only source tree
When an out-of-tree build with read-only source tree was attempted
ZMK would ask gcc to generate dependency information by passing the
option -MMD, which works fine in in-tree builds as the dependency
files are placed alongside the source.

When the source tree is read-only this would obviously fail. To address
that, conditionally pass -MF to put dependency data alongside object files

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>