Fix builds that explicitly use $(ZMK.OutOfTreeSourcePath)
When a project uses out-of-tree source path explicitly, for example
to use it in conjunction with $(wildcard), current logic for out-of-tree
builds would not work correctly. There are several related issues
but the new logic seems to de-compose everything correctly now.

The main fix is the creation of $1.sources which is just $1.Sources
with the out-of-tree source prefix stripped. Subsequently, all objects
are defined with the stripped path.

As a small optimization, the order-only dependency on the directory that
contains object files automatically strips the trailing slash, avoiding
extra rules that cost some time.

Lastly fix typo in internal variable name, $1.sourcesObjC now capitalizes
the trailing C correctly.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>