Copy fixes from Library.So to Library.DyLib
This mainly provides the .alias fix, where we no longer create "naked" files
without .dylib extension if the original library did not carry a version of its
own. In addition a lot of the alias handling and various edge cases were
improved to be easier to read and to be more diff-friendly between the two
library templates.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>