Update out-of-tree vars when loading config.*.mk
ZMK contains several variables that track the configuration of out-of-tree
building. Those are ZMK.IsOutOfTreeBuild, which is useful for make $(if ...)
function, as well as the ZMK.OutOfTreeSourcePath, which is useful as a default
prefix to some paths as it expands to either the empty string or to a path of
the out of tree sources, with the guaranteed trailing slash.

Normally those two are set in z.mk. When the Configure module is imported
zmk further imports the per-project configuration file, config.$(Project.Name).mk
which usually sets ZMK.SrcDir.

Because both ZMK.IsOutOfTreeBuild and ZMK.OutOfTreeSourcePath are defined using
the non-functional if logic, they become stale at that point. This was a long
standing bug that was worked around in other projects by accident, so it was
not surfaced anywhere where out-of-tree builds are default.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>