Add support for using sysroot
The configure script now accepts autotools compatible --with-libtool-sysroot
and then drives the Configure and Toolchain modules to remember the preference.
All rules that invoke the compiler now explicitly pass --sysroot= when the
toolchain sysroot setting is non-empty.

This approach is somewhat fiddly but I wasn't able to find any single good
standard variable that gets passed in all the circumstances and does not result
in duplication, for example, when both CXXFLAGS and CFLAGS are used together.

This allows using zmk with bitbake, through the autotools class, and produce
correct results.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>