Retain customized CC and CXX from configure
When the configure script is invoked with different --build and --host
and cross compiling mode is enabled, do not clobber CC and CXX that is
provided by the configure environment.

This effectively only sets the cross-compiling names of CC and CXX if
they are not already defined.

This matters because poky passes extensive set of flags via CC and CXX,
so forgetting those can cause compile failures. For example for the
Avenger 96 development board, poky uses a soft-float compiler but
switches to hard-float with an argument embedded in CXX:

    CXX=arm-poky-linux-gnueabi-g++  -mthumb -mfpu=neon-vfpv4 -mfloat-abi=hard

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>