Fix Symlink sub-directory install bug
I've noticed that the following rule has a surprising result:

    subdir/name.InstallDir = $(bindir)
    subdir/name.SymlinkTarget=target
    $(eval $(call ZMK.Expand,Symlink,subdir/name))

The symlink, as created by the install target is actually

    $(bindir)/subdir/name -> target

This is both unexpected for the users and problematic on the bigger
design vision, where the makefile could move from one directory to
another without impacting what the results are, as long as some prefix
directory is added or removed accordingly.

I've started looking at fixing this in a backwards compatible way, that
is, to allow developers to opt into the arguably correct behavior while
retaining the broken behavior by default. I came to the conclusion that
it is not the right thing to do because it's an unique property of the
Symlink template. The Program template does not behave this way.  In
fact, nothing else does, since everything except for the symlink uses
the InstallUninstall template under the hood.

In the end both the .sourceDir and .targetDir variables are removed. The
logic, when compared to InstallUninstall, is now very similar.

Fixes: https://github.com/zyga/zmk/issues/80
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>