Add support for OBJC environment variable
The OBJC environment variable is supposed to be like CC but for
Objective-C code. I caught this while looking at the definition of
COMPILE.m in make source code, it refers to $OBJC, and then compiling
with clang++ without gcc/gobjc support on the host.

Adjust XFAIL test to properly pass.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>